+++ src/xyz/oihana/schema/constants/traits/auth/PolicyTrait.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\auth;
+
+/**
+ * The enumeration of all Policy properties.
+ *
+ * @package oihana\schema\constants\traits\auth
+ * @author  Marc Alcaraz
+ */
+trait PolicyTrait
+{
+    const string APPLICATIONS       = 'applications' ;
+    const string APPLICATIONS_COUNT = 'applicationsCount' ;
+    const string COLOR              = 'color' ;
+    const string PERMISSIONS        = 'permissions' ;
+    const string PERMISSIONS_COUNT  = 'permissionsCount' ;
+    const string PROTECTED          = 'protected' ;
+    const string ROLES              = 'roles' ;
+    const string ROLES_COUNT        = 'rolesCount' ;
+    const string SYSTEM             = 'system' ;
+}