+++ tests/xyz/oihana/schema/constants/traits/AuthTraitTest.php
@@ -5,9 +5,37 @@ namespace tests\xyz\oihana\schema\constants\traits ;
 use PHPUnit\Framework\TestCase;
 
 use xyz\oihana\schema\constants\traits\AuthTrait;
+use xyz\oihana\schema\constants\traits\auth\ApplicationTemplateTrait;
+use xyz\oihana\schema\constants\traits\auth\ApplicationTrait;
+use xyz\oihana\schema\constants\traits\auth\InvitationTrait;
+use xyz\oihana\schema\constants\traits\auth\PermissionTrait;
+use xyz\oihana\schema\constants\traits\auth\PolicyTrait;
+use xyz\oihana\schema\constants\traits\auth\RoleTrait;
+use xyz\oihana\schema\constants\traits\auth\ScopeTrait;
+use xyz\oihana\schema\constants\traits\auth\SessionTrait;
+use xyz\oihana\schema\constants\traits\auth\UserTrait;
+use xyz\oihana\schema\constants\traits\auth\WebAPITrait;
+use xyz\oihana\schema\constants\traits\auth\WebApplicationTrait;
 
 class AuthTraitTest extends TestCase
 {
+    public function testComposition(): void
+    {
+        $traits = class_uses( AuthTrait::class );
+
+        $this->assertContains( ApplicationTemplateTrait::class , $traits );
+        $this->assertContains( ApplicationTrait::class         , $traits );
+        $this->assertContains( InvitationTrait::class          , $traits );
+        $this->assertContains( PermissionTrait::class          , $traits );
+        $this->assertContains( PolicyTrait::class              , $traits );
+        $this->assertContains( RoleTrait::class                , $traits );
+        $this->assertContains( ScopeTrait::class               , $traits );
+        $this->assertContains( SessionTrait::class             , $traits );
+        $this->assertContains( UserTrait::class                , $traits );
+        $this->assertContains( WebAPITrait::class              , $traits );
+        $this->assertContains( WebApplicationTrait::class      , $traits );
+    }
+
     public function testComposesAllAuthTraits(): void
     {
         $host = new class { use AuthTrait; };